requirements.txt
# Web framework
fastapi
uvicorn[standard]
python-multipart
jinja2

# PDF processing
pymupdf
Pillow

# Word document processing
python-docx

# Image processing / visual diff
opencv-python-headless
numpy
scikit-image

# AI / LLM
openai>=1.3.0
python-dotenv

# PDF report generation
reportlab

# Auth
authlib
itsdangerous
httpx
